// packages/cosmic/src/vote/interface.ts
/*
 * Copyright (c) Baidu, Inc. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface VoteProps {
  appearance?: '1v1' | 'horizontal' | 'vertical';
  options: Option[];
  targetName?: keyof Target;
  target?: Option | null;
}

export interface Option {
  target: Target;
  value?: number;
  text?: string;
}

export interface Target {
  name: string;
  image?: string;
}

export type VoteData = Required<VoteProps>;

export interface VoteEvents {
  'option-click': Option['target'][keyof Target];
}